1.0.0 release
This is it, version 1.0.0 of `purview` is ready for the masses! This has
been a long time coming and I couldn't be more excited!

Now that we're off the ground, the upcoming commits will [likely] be
heavily focused on adding automated testing and squashing bugs as we
find them.

Please file bugs/feature-requests and submit pull-requests! Let's make
this project great!